Se removieron las llaves foraneas

+++ historias/database/migrations/2024_02_20_153122_create_likes_table.php
@@ -14,8 +14,8 @@ return new class extends Migration
     {
         Schema::create('likes', function (Blueprint $table) {
             $table->id();
-            $table->unsignedBigInteger('history_id');
-            $table->unsignedBigInteger('user_id');
+            $table->integer('history_id');
+            $table->integer('user_id');
             $table->timestamps();
         });
